+++ CHANGELOG.md
@@ -13,3 +13,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Sample schema now pulls in users local time zone rather than defaulting to America/New_York
 - Tested against the 2.1.0 release of Swashbuckle.AspNetCore
+
+## [1.1.0] - 2018-02-25
+### Added
+- Format now matches RFC3339 types when possible https://www.ietf.org/rfc/rfc3339.txt
+	Instant: date-time
+	LocalDate: full-date
+	LocalTime: partial-time (Partial due to lack of offset)
+	Offset: time-numoffset (UTC offset serializes as "+00" and not "Z")
+	OffsetDateTime: full-date